Implement Galaxy profiles.
Specify ``--profile <name>`` with serve, test, or their shed counterparts to maintain a Galaxy database and set of shed installs across invocations.

+++ planemo/io.py
@@ -75,7 +75,10 @@ def shell_join(*args):
     return "; ".join([c for c in args if c])
 
 
-def write_file(path, content):
+def write_file(path, content, force=True):
+    if os.path.exists(path) and not force:
+        return
+
     with open(path, "w") as f:
         f.write(content)
